Add files via upload
I am trying to build a simple upgradeable cheap replaceable module to switch relays for inverters, compressors, and fridges. Ideally the inverter would be off once the battery goes below a level and the compressors would start on a priority set by a pi but this needs to continue working while the pi is offline so I realise I could monitor and control fine with the pi but I would prefer not the rely of any microcontroller for something I should manage solid state. a schmitt trigger set to switch on the inverter at 20v and switch on the other accessories according to priority and time lag ie. the inverter should run and be functioning normal when the fridge turns on. It would be good if a red ark product would do it.